컴포넌트 분리 & 심화과제 해시 라우팅
