+++ src/state.js
@@ -1,0 +1,3 @@
+export const state = {
+  isLogin: false,
+};
